[web] Add a custom layout component
It is the starting point for setting up the proposed layout[1][2] and it
is using none PatterFly component in an attempt to keep it tiny and simple.

Of course, components like Masthead[3], Panel[4], or even [Page] can be
replace some parts of it in the future, if really needed.

About CSS, it tries to follow the same methodology than PatternFly,
i.e., BEM[6].

[1] https://trello.com/c/UBTn9amG (internal link)
[2] https://yast.opensuse.org/blog/2022-02-23/start-2022#dinstaller
[3] https://www.patternfly.org/v4/components/masthead
[4] https://www.patternfly.org/v4/components/panel
[5] https://www.patternfly.org/v4/components/page
[6] http://getbem.com/