[web] Try to not use "link buttons"
Since EOS recommends not making buttons appears as links, we have
changed almost all "logical" buttons from "link" to "secondary" variant.
However, those that are really links (quick actions or shortcuts from
the summary) still using the "link" variant.

There is another special case: the "Do not foo" actions shown in the
footer of some dialogs. By now, we consider they as not destructive
actions in the context of the installer and they do not have the same
"weight" than "Confimr" and "Cancel" actions.