[web] Bring back a removed margin for links at modal footer
To avoid a visual problem when the **button** (which looks like a link
on purpose, to give it "less weight") gets the focus.

Most probably, this action will be inside a section in the future,
instead of in the modal shown for quick changes. Meanwhile, we prefer
having the perceived extra separation when the link is not focused
rather than missing the normal one when the button gets the focus.