[service] QuestionsManager will #wait for a specific list of Questions
Both
- DInstaller::QuestionsManager
- DInstaller::DBus::Clients::QuestionsManager

will #wait for the questions passed as argument.

For the client case this is rather academic as the #ask interface
supports only one question at a time.
But in the main service we could be asking a question while another
service is waiting for an answer to another, and that is not our business.