[service] At boot time, start only the manager
It is not even necessary to start the manager at boot time because it would be
cativated by D-Bus but it reduces the initial response time.

Confused by two kinds of .service files? Welcome to the club!

In a D-Bus .service file: SystemdService=bar.service means: if systemd is
running, delegate starting the service to it (presumably because it is
specialized in managing services and having fine tuned controls for that).
See https://dbus.freedesktop.org/doc/dbus-specification.html#message-bus-starting-services-systemd

In a systemd .service file: BusName=org.foo.bar (implying Type=dbus) means:
the service is considered active when that name is acquired, and systemd will
terminate it when the name is released. See
https://www.freedesktop.org/software/systemd/man/systemd.service.html#Type=