[service] D-Bus policy: don't let non-root even Introspect us
It might seem courteous to allow non-root at least introspect our API
via d-feet, since the documentation is public anyway and this would be
a guaranteed up-to-date version, right?

But our services do a lot of probing at startup, so we might have a bug
where we would modify something as root when triggered by a curious
introspector. Better not allow it.

Let them `sudo -E d-feet`.