[web] Fix fonts import
Since we're using an specific rule in webpack.config.js file for fonts, we
can't rely on @use Sass at-rule for importing the fonts.scss file. Instead, we
have to import it in the index.js file directly, for making webpack aware of
it.

What is more, most probably the fonts.scss file could be renamed to fonts.css
since it does not contain nothing to process with Sass.

Related to https://github.com/yast/d-installer/pull/385