[web] Update aliases for using ~/ instead of @
Because "@" is used by npm for packages published under a named scope[1]. From
their documentation:

  > The scope name is everything between the @ and the slash

Additionally, "~" use to kind of means "home" or "root", which suits well for
an alias pointing to the "root" directory.

Consequently, it is easier to understand now whether an import comes from a
dependency or the source code of the project.

[1] https://docs.npmjs.com/about-scopes