First skeleton of Manager1 API docs
Seed it:

    sudo busctl \-\-verbose \
      \-\-address unix:path=/run/d-installer/bus \
      call \
       org.opensuse.DInstaller \
      /org/opensuse/DInstaller/Manager1 \
       org.freedesktop.DBus.Introspectable \
       Introspect

(the escaped dash is so that a double dash does not invalidate a XML
comment, yay)
Remove other interfaces, leave only one.

Render it:

    for f in  org.opensuse.DInstaller.Manager1.xml; do
      docbook=${f%.xml}.ref.xml
      ./spec-to-docbook.sh spec-to-docbook.xsl $f $docbook
      xmlto --skip-validation html-nochunks $docbook
    done

TODO:

Read this with the library and validate that it matches (does not
conflict) with what the code declares.
Or maybe generate stub code?