[web] Drop Sidebar.OpenButton component
Because the use case [1] for which it was introduced was already gone:
to provide a link for directly opening the sidebar from a text teaching
the user where to find page related actions.

We can bring back the component if needed, but let's get rid of it now.

This commit "reverts" f1a56eba9, df85b3c70, and a bit of f39f65dc1

[1] https://github.com/openSUSE/agama/pull/500