[web] Add a the initial version of a notification mark
A simple component for rendering a notification/status mark, consisting
of an absolute right positioned circle on top of a relative positioned
element.

The idea is to catch the users' attention for clicking on the element
presenting the mark, but having in mind that they can continue with the
installation without doing so.