[web] Increase the z-index of the sidebar
To avoid other elements overlapping it. A very high value was chosen
because PatternFly uses values from 100 to 600 at the time of writing.
See [1]

Another option could be to use the PatternFly highest value, but for now
it does not seem necessary to couple the Sidebar z-index to a PatternFly
variable.

[1] https://github.com/patternfly/patternfly/blob/7ddc2de5cfb96f204e62730878a9e77bed64b998/src/patternfly/sass-utilities/scss-variables.scss#L154-L160