[web] Ensure Sidebar siblings don't remain inert forever
While unmounting the Sidebar is unlikely to happen, it's better to make
the component a good citizen by "restoring" its siblings by removing the
`inert` and `aria-hidden` attributes. That way, we're minimizing the
risk of users being trapped in an issue similar to the one described at
https://github.com/patternfly/patternfly-react/pull/9096

Please note that at the time of writing it's not a real restore because
the component isn't tracking the status of these attributes BEFORE
adding them. However, removing the supposedly added attributes seems
enough for now.