[web] Users: improve the root auth section
Providing more comprehensive information to the user if there is no
authentication method for the root user, which may result in a system
without administrative access.

Once a method is defined, a table listing authentication methods will be
displayed instead. This might be changed in a near future for a better
alternative.