[web] Remove layout slot no longer needed
At https://github.com/openSUSE/agama/pull/781, the Sidebar was created
directly from the Layout by using a slot for teleporting its content
from outside. That was needed because the Layout was moved to a
provider. Later, https://github.com/openSUSE/agama/pull/796 restored the
previous behavior of mounting both, the Sidebar and the Layout, from
src/App.jsx instead of having "content components" in the providers,
which was  weird if not wrong.

However, the slot created for the Sidebar content was not removed and it
was creating an empty div node between the sidebar and the main content.
Although it wasn't producing any visible side effect in the UI, it has
been removed since it is actually no longer needed.