[web] Drop not used Section#openDialog prop
It was there almost from the beginning for allowing allowing a modal
dialog to be opened directly from a section title. However, as the
installer has evolved, we have limited the title behavior to navigation
only which is more expected and less confusing. If a section needs to
open a dialog it's better to use an explicit option inside the section
content instead of hiding these options behind the section header.

We can reintroduce the feature in the future if really needed, but
maintaining it just in case isn't an option.