[web] Change how storage selectors are rendered
To make it possible reusing the same code for rendering either, a list
or a selector. There is room for improvements yet, but it should be
enough for now to avoid having list marked as "listbox" with elements
set as "role=option".

This change also proposes a different way to style them, avoiding
setting a grid by default for each type of selector since it's not
always the case. E.g., the space policy selector has a so simple layout.
In any case, this is a subject for changes in the short term since
selectors must be reworked ASAP.