[web] Reduce the use of data-type attribute
They were used for styling some elements that can be styled in a
different way. In any case, we still needing something to make these div
nodes accessible since now we (sigth people) figure out what they are
because we're able to see them well distributed in the UI. But I'm not
sure what a mess it can be if something read all these information
aloud.