[web] Make possible opening the Sidebar via prop
In the context of improving the core/Page component and how the Agama
layout is handled, it's needed to have a way of opening the Sidebar from
outside instead of keeping it responsible of rendering a button for such
an action.

Using a prop to do so following the recomendation found in the
"Pitfall" warning of "Exposing your own imperaive methods" section in
the React#useImpreativeHandle documentation. See [1]

Please, note that this commit also drops the logic for displaying a
"notification mark" because there is already a work in progress for
getting ride of such a feature once we have decided that the sidebar is
for placing Agama stuff only. See [2]

[1] https://react.dev/reference/react/useImperativeHandle#exposing-your-own-imperative-methods
[2] https://github.com/openSUSE/agama/pull/886