[web] Improve page and layout management
Now that everything in Agama is on a page, the core/Page component
should take care of layout. This can be considered a first step toward
dropping both the layout/Layout component and react-teleporter
dependency.

In summary, the goal of stopping using react-teleporter dependency is to
improve DX by writing JSX closer to final HTML. Also, stopping using
React Portals behind the scene for placing content will help now that
actually it's not needed to influence the layout from Page children.

When the user navigates from one page to another, mounting and rendering
the layout is somehow expected and shouldn't harm performance noticeable.

Creating new Pages is now much easier without having to import slots for
titles, icons, actions, etc. It will be sufficient to use the Page props
and Page subcomponents.