[web] Drop react-teleporter dependency
Due to a switch from layout slots to a page component that places
everything in the right place.

It has nothing to do with react-teleporter itself, which is a useful and
tiny dependency that works as expected. It's simply because Agama has
evolved, and it no longer needs to influence the layout from its
children. Instead, the core/Page component has become a central
component, and all content within the app must be contained on a page,
along with the sidebar having become a static area for Agama-related
information.