[web] Move "static" sidebar content to the component
The Sidebar holds the installer options and they are always rendered as
its children, so mounting them from the component mounting the Sidebar
makes no sense. They've instead been moved to the Sidebar component
itself, but with the ability to add new items as children just in case
such a feature might be useful.