[web] Rename PageMenu and aliases it in Page
The former name "PageOptions" was not so accurated probably becuase it
dates back to when these entries where placed as a "ContextualActions"
in core/Sidebar. Later, they were moved to the header as a "PageOptions"
to avoid collision with the "PageActions" but clearly "PageMenu" is a
better name for them at this moment.