[web] UserPage: start using HTML <strong> element
Before this change, the CSS class bold was used for emphasizing the "No
X defined" yet message. However, important messages are actually those
telling the user something has to be defined.

Thus, this commit uses the HTML <strong> element to wrap them and
deliver a more semantic document instead of relying on the CSS 'bold'
class.