[web] Do not use core/Section for list of issues
Since core/Section is reserving space for icons, it looks a bit weird
when used inside a dialog for displaying a list of items. Too much space
on the left.

Thus, stop using it by now until the component allows conditionally
reserve such an space or not depending on whatever, a prop or parent.

Also, start using a plain HTML <ul> element instead of the PF/HelperText
for wrapping the list of issues.