[web] Storage: mount point selector shown only if use can change it
Problem :

When editing a file system, the form shows a disabled selector for the
mount point field. Using a disabled widget does not make sense in this
case because the mount point will never be editable. The disabled widget
should be replaced by plain text.

Moreover, the selector currently uses a FormSelect component. It should
be replaced by the Select component, similar to the file system type
selector.

Solution :

- Updated the MountPointFormSelector to use Select component Instead of
the previous FormSelect component.

- The select component adds more customizability and some advanced
features as well.

- Updated the mountpoint field to show the mountpath when currentVolume
is defined and prompts the user to choose the mountpoint from the list
of volume templates in case the currentVolume is undefined.

Testing :

Tested manually

Screenshots:

Before:
![Screenshot 2024-01-09
232254](https://github.com/openSUSE/agama/assets/125825782/a68fb82c-977a-447d-a54a-f2b673eb7da1)
![Screenshot 2024-01-09
232710](https://github.com/openSUSE/agama/assets/125825782/a8960c3e-a437-4e03-946f-d8704da13a63)

After:

![image](https://github.com/openSUSE/agama/assets/125825782/21680bfa-da97-4e93-a0b2-692fcf42f71b)


![image](https://github.com/openSUSE/agama/assets/125825782/17ed2255-1ec9-4267-a0bd-6ef531059b81)

---------

Co-authored-by: David Díaz González <dgonzalez@suse.de>