[web][doc] Change references to JSDoc (#1009)
A few weeks ago, the [documentation output was reworked a
bit](https://github.com/openSUSE/agama/pull/976) to be ready for a near
future in which all the project documentation could be "managed" by a
static site generator. These adjustments implied a few changes that were
not reflected in the README file.