[web] Adds a core/Selector component
Despite being in a preliminary state, it is a good start to unify all
the selectors. Also, it is the first step toward switching from the
Listbox[1] to Grid[2] pattern, because selector options now include at
least one interactive element: a radio button when mounted as a single
selector or a checkbox when multiple selection is allowed.

Grid pattern implementation (or its adoption, if finally the project
ends up using a third-party dependency) will take a long time, but sadly
it isn't expected to be completed in the near future[3].

Hence, this commit is part of a request to centralize selectors base
code and be ready for better implementation when time and priorities
permit.

[1] https://www.w3.org/WAI/ARIA/apg/patterns/listbox/
[2] https://www.w3.org/WAI/ARIA/apg/patterns/grid/
[3] https://github.com/openSUSE/agama/discussions/786#discussioncomment-8209053