[web] Adapt storage/DeviceSelector to core/Selector
Please, be aware that changes sent in this commit are not intended to be
permanent since almost the whole involved logic will be reworked to
accomplish what is stated https://github.com/openSUSE/agama/pull/1031.

Thus, there are things that could be done better but postponed on
purpose to be addressed in another request.