[web] Add method to reset localStorage in tests
Sometimes could be needed to reset the window.localStorage cache between
test examples, as it was the case for recently added tests for the space
policy section.

Thus, instead of having such utility hidden in and repeated in needed
tests, it has been moved to a test-utils.js file where, with a bit of
luck it will be more discoverable for others.

Additionally, that new helpers allows to set an initial cache too by
passing it an object with a collection of items compatible with Web
Storage API setItem method.