web: add method to retrieve content from an object
It allows providing a collection of paths to look for in the given
object, returning the first content found.

Although it could have a better name and/or signature (initially
`findContent({ in, at })`), the motivation for adding it arose in the
context of creating an expandable selector for the storage proposal.

Such a selector will receive a collection of device objects to be
represented in an expandable table. These elements might have children,
which will be hidden when their parent is collapsed. But not all kinds
of devices will have these children at the same key/level/path. At the
time of writing, we know that partitions of a disk can be found at
`partitionsTable.partitions` location. But that might not be the case
for LVM VGs, which most probably will be at `vgs` or `volumes` path.

Thus, relying on the method added in this commit, the selector will be
ready to accept a path or paths to search for children. In addition to
making it more versatile, it eliminates the need to create separate
collections for reallocating children of each kind of item.