web: stop keeping selected items in the selector state
It's not needed in the current Agama forms workflow, in which the form
always hold the state.

This is something that maybe can be changed in the future to avoid
having widgets that are re-rendered all the time without no other reason
than the "requirement" of lifting their state. Would be more
"reasonable" having forms able able to ask about the data before sending
it.