web: add Result section to Storage page (#1088)
Replace the `Actions` section in the storage proposal for a `Result` one
which presents the proposal result in a more appealing way displaying
how the storage would look after installation instead of just the list
of actions. Among others bits, it hints the user about things like
file-systems are going to be created, resized actions and a subtle
emphasis in deletions, if any.

The full, plain list of actions is now available in a modal dialog
linked right before the new introduced table.