web: Add initial version of a segmented control pattern
Initially based on props only, but most probably it will evolve to
provide subcomponents to make possible to build more detailed options by
just writing JSX.

It has been added to be used in the new device selection dialog that
will land in the Storage page and after discarding plain HTML input
radios and tabs components. The radios were discarded because target
actions are not actually form options but rather "form/mode pickers"
which have an immediate effect on the user interface after selecting
one. The tabs for avoiding mislead the user that they are navigating
between these options. In other words, to ensure users are aware of the
mutually exclusive settings they are in front of.

It is inspired by the segmented control pattern. To read more have a
look to below links

* https://primer.style/components/segmented-control
* https://medium.com/tap-to-dismiss/a-better-segmented-control-9e662de2ef57