[WIP] web: New PartitionsField
This groups in one ExpandableSection the content that was before at
ProposalVolumes, SnapshotsField and BootConfigField.

Tests are not adapted, the test file is simply renamed so far.

Type checking and documentation of the methods is still marked as TODO.

Ideally, this should contain a separate component VolumesField on its
own file instead of incorporating the table directly in the advanced
view. That can be done in the future.