web: Adapt SpacePolicyDialog for using Field/RadioField
Which actually introduced a bit of complexity when it comes to build the
UI and at CSS level. Also, it does not looks well in small screens,
requiring much more "CSS micromanagement" to have it working in a decent
way.

Relying on Field/Radio picker makes it more handy out-of-the-box at the
same time that the goal of OptionsPicker is achieve: the information the
user needs to decide one option or another is already there. And it
looks more closer to native platform.