web: Allow influence Popup size
By ignoring the PF/Modal's `variant` prop and adding `blockSize` and
`inlineSize` to core/Popup.

In addition to "auto", they support "small", "medium", and "large"
values that are converted into CSS classes to set the component size
based on the viewport.

The intention is to keep the dialog at a constant size regardless of its
content.

Developers should decide which size is best for a dialog based on its
content and behavior.