GitHub Action: bind mount /usr/bin/gh into the containers
We use the `gh` GitHub command line tool to create pull requests.
The tool is installed inside the Tumbleweed container using `zypper`
as usually.

The problem is that the installation might fail in some cases
(GPG key problem, network error,...) and this fails whole action.

But we actually do not need to install the tool, it already is pre-installed in
the default Ubuntu host system. And because the `gh` tool is a statically
linked binary without any dependencies we can simply just bind mount it from
the host inside the container without any installation.