Tumbleweed: more sensible min for generic volumes
That min size is only relevant to populate the form in the UI with an
initial default value. Zero is a weird value for that because is not
even a valid min. All the other minimums in the product (/, /home,
swap...) are in the scale of Gigabytes. So 1GiB looks more sensible as
starting point for the user to adjust it.