Use a timeout when calling `setxkbmap`
When calling `setxkbmap` in development the `root` user might
not have access to the `:0` X display.

Maybe there is XDM running or another user is logged in. In that
case `setxkbmap` prints

  Authorization required, but no authorization protocol specified

error message in an infinite loop and gets stucked.

Additionally the $DISPLAY might be already set so we should honor
that value.