fix(web): better layout for transactional alert
The storage alert for a transactional root file system was placed in the
header to keep it sticky there all the time. However, we've concluded
that should be enough to put it at the very top of the page with a
better, more prominent presentation. To so so, it stop using the custom
Agama/Reminder component in favor of PatternFly/Alert.