Reduce repo loading UX latency by preselecting the first product
Repo metadata downloading can take 20 seconds and that is a big delay in user
experience.
Preselecting a product is a simple way to load the data even before the user
logs in.
This is most effective when all products share the same repos (like openSUSE
Tumbleweed and openSUSE MicroOS do)

OTOH now we have preselected a product and taken that important choice away*
*: still there via Change Product, but far less prominent.