fix(web): change element wrapping core/EmptyState children (#1462)
## Problem

The elements in the login page are rendered wrongly since
https://github.com/openSUSE/agama/pull/1441. That's because the `Flex`
component used for wrapping the `core/EmptyState` children. It is
missing the "direction" prop for telling it to layout children
vertically instead of horizontally. Moreover, it looks like a `Stack`
element is a better choice for this case.

## Solution

Adapt `core/EmptyState` for using an `<Stack hasGutter>` instead of a
`<Flex>` configured to behave almost the same.

## Testing

- Tested manually