refactor(web): stop using <Page> as route element
Using it as a wrapper for PF/PageGroup accepting children instead of a
wrapper for a router <Outlet>. The latest is already rendered by the
layout.

Please, not that this change is just the starting point of a bigger Page
component refactor that should help to improve both, the DX experience
and the resulting DOM output.