fix(web): restore radio inputs in product selection
Since it's arguably that they might confuse users and also because of
consistent with the rest of the UI. They could be drop or hidden later,
but thinking globally in all selectors and implications of doing so.

Also, this commit improves a little the proposed layout by making use of
the label for "increasing" the clickable area instead of adding a
handler for the PF/Card component itself. Again, this can be refactored
later if really needed.

Read the reasoning at
https://github.com/openSUSE/agama/pull/1550#pullrequestreview-2269021089
and/or https://github.com/openSUSE/agama/pull/1550#issuecomment-2326622683.