fix(web): do not lost focus when selecting a product
As a result of an internal component defined in the body of the main
component (aka "render function"), the focus was lost each time the main
component got re-rendered because of a "selected product" changed.

This happens because these internal components are actually not
re-rendered but remounting since they were re-defined :/ A bit messy,
but there is a lot of documentation about it. To link a few,

  * https://react.dev/learn/preserving-and-resetting-state#different-components-at-the-same-position-reset-state
    - Which contains a "Pitfall" warning that says: **This is why you
      should not nest component function definitions.**
  * https://www.developerway.com/posts/react-re-renders-guide#part3.1