fix(web): avoid PF/Radio complaints
The ProductSelectionPage component was laying out PF/Radio inputs in a
not expected way, making such a component complaint because missing
prop

> https://github.com/patternfly/patternfly-react/blob/a3ffb39a0cc2c9130f7db86ad3186787ba12648e/packages/react-core/src/components/Radio/Radio.tsx#L61-L64

Thus, the selector should either, to adhere to expected layouts or to
simply use a plain input[type="radio"] which the corresponding CSS class
to look like a PF/Radio. This commit implements the latest.