refactor(web): move Install button to top bar
It implies to remove it, and its ad hoc section, from the Overview page.

As an intermediate step, it also change the presentation of the Overview
page, which only uses the "two columns" layout if there is any issue to
render. Not polished in details because it is subject to future changes
to make the issues list always accessible/visible from any other place.