feature(web): render issues link at header when needed
As a counterpart of the InstallButton, an IssuesLink is displayed when
the installation is not possible to make the user aware why the
InstallButton is not there.

This is a temporary navigation link to the overview page that, most
probably, will be replaced by a toggler for a Notification Drawer
https://www.patternfly.org/components/notification-drawer always
accessible from the top bar.