fix(web): change when root auth page is shown
Before this commit, the RootAuthMethodsPage would prompt users as soon
as there was an issue in the user scope, ensuring that at least one root
authentication method was defined. After reconsidering the behavior, it
has been decided that relying on the available root user data is more
robust. This means the page will now only display if no root
authentication method is defined.

Additionally, an optimistic update has been added at the query level to
prevent unnecessary redirection when the App router is reached before
the query cache is updated after a successful server response.