Fixed firmware cleanup script (#1812)
## Problem

- See #1756
- Some firmware files were missing in the Live ISO

## Solution

There were basically two problems:

1. Few kernel modules use wildcards to specify the firmware files, e.g.
`ath11k/WCN6855/hw2.1/*`. That was not supported by the script, it
expected only individual file names referenced by drivers.
2. Some firmware files use symlinks, e.g. the driver refers to `foo.bin`
file but in some cases it is actually a symlink pointing to
`foo-1.0.bin`. The script then deleted the `foo-1.0.bin` file as it
looked like not used by drivers. The result was a dangling symlink and a
missing firmware file.

Both issues were fixed. The glob patterns are evaluated using the
`Dir.glob` function and if the referenced file is a symlink then it also
keeps the target file where the symlink points to.

## Testing

- Tested manually, the firmware for Atheros cards is now present
- I compared the result with installation-images and now we basically
include the same firmware as installation-images do in the classic
Tumbleweed installation
- The ISO size increased just by ~13MB, fortunately not that much...

---------

Co-authored-by: Imobach González Sosa <igonzalezsosa@suse.com>