fix(web): proper handle netmask
The backend expects a prefix length instead of netmask, but
unfortunately current code base is not ensuring of making the conversion
before sending a request, making the connection form fail silently.

This commit fix it by adapting the formatIp util method for relying in
ipPrefixFor to make such a conversion.