fix(web): update tests looking for actions in rows
To properly look for the menu actions at the document body instead of
the row where the toggler lives. This change is needed because in PFv6
"The default value of appendTo on Dropdown, Select, and Popper has been
updated to `document.body`"

Maybe in the future the appendTo prop is added to the actions menu
instead.